No point in tracking e.g. the .html pages
in the 'docs' folder since these are generated by pdoc and will be regenerated upon every update on the 'release_v1.0.0 branch' through GH Actions and the '.github/workflows/docs.yml' file. The advantage of git untracking everything in the 'docs' folder is that one does not have to run pdoc locally and 'git add' and 'git push' all the .html pages updated by pdoc.