Removing all those lines makes the modules
in the generated html pages by pdoc appear. Previously, although the docstrings for all classes and functions were rendered in html pages, the modules containing those classes and functions were not, i.e. the user reading the documentation was not informed about where those classes and functions could be found.